change linters, add dry-run option

+++ ueil_tagger/client.py
@@ -11,29 +11,31 @@ if TYPE_CHECKING:
     from ueil_tagger.types import Uuid, WebAPIRecord
 
 
-PEOPLE_ENDPOINT="https://actionnetwork.org/api/v2/people"
-TAGS_ENDPOINT="https://actionnetwork.org/api/v2/tags"
+PEOPLE_ENDPOINT = "https://actionnetwork.org/api/v2/people"
+TAGS_ENDPOINT = "https://actionnetwork.org/api/v2/tags"
 
 
 class Client:
     api_key: str
+    read_only: bool
 
-    def __init__(self, api_key: str) -> None:
+    def __init__(self, api_key: str, dry_run: bool = False) -> None:
         self.api_key = api_key
-
+        self.read_only = dry_run
 
     def __delete(self, url: str, background: bool = True) -> bool:
-        headers =  {
+        headers = {
             "api-key": self.api_key
         }
         params = {}
         if background:
             params["background_request"] = "true"
         logging.debug("(DELETE) %s params=(%s)", url, json.dumps(params))
+        if self.read_only:
+            return True
         rs = requests.delete(url, headers=headers, params=params, timeout=10)
         logging.debug("...%s: %s bytes", rs.status_code, len(rs.content))
         return rs.ok
-
 
     def __post(self, url: str, data: Optional[Any],
                background: bool = True) -> bool:
@@ -46,6 +48,8 @@ class Client:
             params["background_request"] = "true"
         logging.debug("(POST) %s params=(%s) data=(%s)",
                       url, json.dumps(params), json.dumps(data))
+        if self.read_only:
+            return True
         rs = requests.post(url, json=data, headers=headers, params=params,
                            timeout=10)
         logging.debug("...%s: %s bytes", rs.status_code, len(rs.content))
@@ -54,12 +58,11 @@ class Client:
                           rs.text)
         return rs.ok
 
-
     def __get(self, url: str,
               params: Optional[dict[str, str]] = None) -> WebAPIRecord:
         if not params:
             params = {}
-        headers =  {
+        headers = {
             "Content-Type": "application/json",
             "OSDI-API-Token": self.api_key
         }
@@ -68,13 +71,11 @@ class Client:
         logging.debug("...%s: %s bytes", rs.status_code, len(rs.content))
         return cast("WebAPIRecord", rs.json())
 
-
     def get_tags(self, page: int = 1) -> WebAPIRecord:
         params = {
             "page": str(page)
         }
         return self.__get(TAGS_ENDPOINT, params)
-
 
     def get_people(self, page: int = 1,
                    modified_since: Optional[datetime] = None) -> WebAPIRecord:
@@ -82,14 +83,13 @@ class Client:
             "page": str(page)
         }
         if modified_since is not None:
-            params["filter"] = f"modified_date gt '{modified_since.isoformat()}'"
+            date_filter = f"modified_date gt '{modified_since.isoformat()}'"
+            params["filter"] = date_filter
         return self.__get(PEOPLE_ENDPOINT, params)
-
 
     def get_person(self, person_uuid: Uuid) -> WebAPIRecord:
         url = f"{PEOPLE_ENDPOINT}/{person_uuid}"
         return self.__get(url, {})
-
 
     def set_tagging_for_person(self, tag_uuid: Uuid, person_uuid: Uuid,
                                background: bool = True) -> bool:
@@ -104,7 +104,6 @@ class Client:
         }
         return self.__post(url, data, background=background)
 
-
     def get_taggings_for_person(self, person_uuid: Uuid,
                                 page: int = 1) -> list[Uuid]:
         url = f"{PEOPLE_ENDPOINT}/{person_uuid}/taggings"
@@ -118,7 +117,6 @@ class Client:
             tag_uuid = tag_href.split("/")[-1]
             tag_uuids.append(tag_uuid)
         return tag_uuids
-
 
     def delete_tagging_for_person(self, tag_uuid: Uuid, person_uuid: Uuid,
                                   background: bool = True) -> bool:
